fixed6 added clean path

+++ backend/src/services/image.service.go
@@ -1,6 +1,7 @@
 package services
 
 import (
+	"errors"
 	"fmt"
 	"io"
 	"mime/multipart"
@@ -22,9 +23,17 @@ func NewImageService() *ImageService {
 
 // Сохранение изображения
 func (s *ImageService) SaveImage(folder string, file multipart.File, fileHeader *multipart.FileHeader) (string, error) {
-	// Создаем папку, если она не существует
+	// Создаем безопасный путь
 	savePath := filepath.Join(s.BasePath, folder)
-	err := os.MkdirAll(savePath, os.ModePerm)
+	savePath = filepath.Clean(savePath)
+
+	// Проверяем, что путь находится внутри BasePath
+	if !strings.HasPrefix(savePath, s.BasePath) {
+		return "", errors.New("invalid folder path")
+	}
+
+	// Создаем папку, если она не существует
+	err := os.MkdirAll(savePath, 0750)
 	if err != nil {
 		return "", fmt.Errorf("failed to create directory: %w", err)
 	}
@@ -41,6 +50,14 @@ func (s *ImageService) SaveImage(folder string, file multipart.File, fileHeader 
 		// Если файл существует, добавляем префикс времени, чтобы избежать конфликта имен
 		newName := fmt.Sprintf("%d_%s", time.Now().Unix(), cleanName)
 		filePath = filepath.Join(savePath, newName)
+	}
+
+	// Очистка пути для безопасности
+	filePath = filepath.Clean(filePath)
+
+	// Проверяем, что путь находится внутри BasePath
+	if !strings.HasPrefix(filePath, s.BasePath) {
+		return "", errors.New("invalid file path")
 	}
 
 	// Открываем файл для записи
@@ -65,7 +82,12 @@ func (s *ImageService) SaveImage(folder string, file multipart.File, fileHeader 
 // Удаление изображения
 func (*ImageService) DeleteImage(filePath string) error {
 	// Преобразуем путь в формат ОС
-	fullPath := filepath.FromSlash(filePath)
+	fullPath := filepath.Clean(filepath.FromSlash(filePath))
+
+	// Проверяем, что путь безопасен
+	if !strings.HasPrefix(fullPath, "src/public") {
+		return errors.New("invalid file path")
+	}
 
 	// Удаляем файл
 	err := os.Remove(fullPath)
